backend/requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dgl==0.9.1
docopt==0.6.2
fastapi==0.87.0
filelock==3.8.0
h11==0.14.0
huggingface-hub==0.10.1
idna==3.4
joblib==1.2.0
networkx==2.8.7
nltk==3.7
numpy==1.23.4
packaging==21.3
pipreqs==0.4.11
psutil==5.9.3
pydantic==1.10.2
pyparsing==3.0.9
pytorch-ignite==0.4.10
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
sniffio==1.3.0
starlette==0.21.0
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
yarg==0.1.9
pycryptodome==3.15.0